Fix warnings when building with Visual Studio 2010
Use new CTK tag which corrects the warnings from
ctkDependencyGraph && ctkErrorLogFDMessageHandler.
Disable crt secure warnings: -D_CRT_SECURE_NO_WARNINGS
Disable warnings from VTK using the CTestCustom

+++ Libs/VTK/Widgets/Testing/Cpp/msvVTKProp3DButtonRepresentationTest1.cxx
@@ -166,7 +166,7 @@ int msvVTKProp3DButtonRepresentationTest1(int, char* [])
 
   // Compute Rendering on the representation when following the camera
   prop3DButtonRep->FollowCameraOn();
-  if (prop3DButtonRep->GetFollowCamera() != true)
+  if (prop3DButtonRep->GetFollowCamera() != static_cast<int>(true))
     {
     std::cerr << "Error: FollowCameraOn method is not effective" << std::endl;
     return EXIT_FAILURE;
